% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataSummary.R
\name{DataSummary}
\alias{DataSummary}
\title{Summarize Data}
\usage{
DataSummary(Y, X, XC, data)
}
\arguments{
\item{Y}{The name of the variable to be predicted}

\item{X}{Name of the quantitative independent variable.}

\item{XC}{Name of the qualitative dependent variable.}

\item{data}{Data set to be used for the analysis.}
}
\value{
This function will return 2 plots (in one window). There is a plot of
   The data with the regression lines added and a bar plot of each group of the
   categorical data. The following are also returned in a named list:
   \itemize{
      \item{betas}{Betas produced from a linear model of the data}
      \item{CI}{Confidence intervals for beta estimates of a model}
      \item{lmsum}{Summary output of a model used for analysis}
      \item{datasum}{A summary of the entire data set}
      \item{groupSum}{A summary of the data in each group}
   }
}
\description{
This function will look at a data set and the names of variables of the
   data to conduct some summary output and helpful plots to better examine the data.
   This is helpful to see if there are any issues with a linear model for this specific
   data. Note that a red line signifies the base level on the plot.
}
\examples{
print(DataSummary("CH4", "TIME", "VHA-Added", SLUDGE))
}
