% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResAn.R
\name{ResAn}
\alias{ResAn}
\title{Analysis of Residuals}
\usage{
ResAn(model, data)
}
\arguments{
\item{model}{The model to be examined, of the form y ~ x1 + x2}

\item{data}{The data frame of which the data is pulled from}
}
\value{
This function will return 2 plots (in one window). There is a plot of
   residuals vs predicted, and a plot of cooks distance. The
   following are also returned in a named list:
   \itemize{
      \item{InfMeasures}{A dataframe that contains the indicies and reasons for which
      data values could be considered as having high influence}
      \item{anova}{A summary output from an anova test on the model}
      \item{resids}{A vector of residuals from the model}
      \item{P}{The Hat matrix for the model}
      \item{r.sqd}{The calculated \eqn{r^2} value of the model, supplied by \code{lm}}
      \item{AIC}{The calculated AIC value of the model}
   }
}
\description{
This funciton looks at the model passed in and will conduct various
   reports about the data that can be helpful, specifically we are looking at potential
   issues with the residuals and if any specific points should be examined further and
   for which reason.
}
\examples{
print(ResAn(TIME ~ NUMBER + BROWSER, BROWSER))
}
